backend/requirements.txt
fastapi==0.104.1
httpx==0.25.2
aiofiles==23.2.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

pyyaml==6.0.1
python-dotenv==1.0.0

langchain-openai==0.3.33
openai==1.107.2
anthropic==0.7.7
tiktoken==0.11.0

pydantic==2.11.9
email-validator==2.1.0

python-dateutil==2.8.2

nltk==3.8.1
textstat==0.7.3

python-json-logger==2.0.7
rich==13.7.0
httpx==0.25.2

aiosqlite==0.19.0
uvicorn==0.35.0
pyjwt==2.10.1